Split synadm.py into multiple modules and packages
