+++ cli
@@ -1,0 +1,9 @@
+#!/usr/bin/env python3
+
+from synadm import cli
+from synadm.cli import user, room
+
+
+if __name__ == "__main__":
+    cli.create_config_dir()
+    cli.root(obj={})
